* move assets into `assets.h` * use standard C rand() for fruit placement, the Z80 ASM is a bit advanced
