+++ examples/snake/src/title.h
@@ -1,0 +1,2 @@
+void play_title(void);
+void hide_title(void);
